feat: add credential claim verification
